requirements.txt
fastapi
uvicorn
transformers
requests
PyPDF2
python-docx
llama-index
torch
qdrant-client
langchain
pandas
langchain_google_genai